Create shared library for synteny functions, abstract
