+++ src/js/annotations/synteny-collinear.js
@@ -1,4 +1,5 @@
 import {d3} from '../lib';
+import {getRegionsR1AndR2} from './synteny-lib';
 
 function writeSyntenicRegion(syntenies, regionID, ideo) {
   return syntenies.append('g')
@@ -21,26 +22,6 @@ function writeSyntenicRegion(syntenies, regionID, ideo) {
       d3.selectAll(ideo.selector + ' .syntenicRegion')
         .classed('ghost', false);
     });
-}
-
-function getRegionsR1AndR2(regions, ideo) {
-  var r1, r2;
-
-  r1 = regions.r1;
-  r2 = regions.r2;
-
-  var r1ChrDom = document.querySelector('#' + r1.chr.id + '-chromosome-set');
-  var r1GenomeOffset = r1ChrDom.getCTM().f;
-  var r2ChrDom = document.querySelector('#' + r2.chr.id + '-chromosome-set');
-  // var r2GenomeOffset = r2ChrDom.getBoundingClientRect().top;
-  var r2GenomeOffset = r2ChrDom.getCTM().f;
-
-  r1.startPx = ideo.convertBpToPx(r1.chr, r1.start) + r1GenomeOffset - 12;
-  r1.stopPx = ideo.convertBpToPx(r1.chr, r1.stop) + r1GenomeOffset - 12;
-  r2.startPx = ideo.convertBpToPx(r2.chr, r2.start) + r2GenomeOffset - 12;
-  r2.stopPx = ideo.convertBpToPx(r2.chr, r2.stop) + r2GenomeOffset - 12;
-
-  return [r1, r2];
 }
 
 function writeSyntenicRegionPolygons(syntenicRegion, x1, x2, r1, r2, regions) {
